expose statsigmetadata (#425)
### TL;DR

expose `StatsigMetadata`​for iOS.

### Why make this change?

Feature request from a customer - they'd like runtime access to this data on mobile platforms.

Exposing `DeviceEvironment`​was considered, but would have taken a decent chunk of refactoring work, and it doesn't include all the relevant fields - `StatsigClient.currentUser.deviceEnvironment`​ is the actual source of truth.

Returning a copy of the `currentUser.deviceEnvironment`​dict was considered, but is not ideal from a usability standpoint - it might be viable if the backing keys were static constants and also exposed.

### What changed?

- Added a `StatsigMetadata`​struct to pass to users of the SDK
    - Includes an internal  helper for populating it from `currentUser.deviceEnvironment`​
- Added `getStatsigMetadata()`​to `Statsig`​and `StatsigClient`​
- Migrated relevant raw strings keys to static constants